W1-06 - Add Pinia Store - Update Unit test & code coverage
