+++ .github/workflows/build.yml
@@ -31,12 +31,12 @@ jobs:
       - name: Lint
         run: npm run lint --if-present
 
-      # - name: Unit tests with coverage
-      #   run: npm run test --if-present
+      - name: Unit tests with coverage
+        run: npm run test --if-present
 
-      # - name: List coverage artifacts
-      #   if: always()
-      #   run: ls -lah coverage || true
+      - name: List coverage artifacts
+        if: always()
+        run: ls -lah coverage || true
 
       - name: Build
         run: npm run build --if-present
