+++ .gitignore
@@ -8,10 +8,10 @@ dist
 
 # Coverage & static analysis
 coverage/
+app/tests/
 .scannerwork/
 .sonarlint/
 .eslintcache
-
 # Node dependencies
 node_modules
