--- tests/smoke.spec.ts
@@ -1,4 +1,0 @@
-import { describe, it, expect } from 'vitest'
-describe('smoke', () => {
-  it('works', () => expect(1 + 1).toBe(2))
-})
